requirements.txt
django == 4.1
djangorestframework == 3.12.4
requests == 2.28.1
python-decouple == 3.6
django_cors_headers == 3.13.0
djangorestframework-simplejwt == 5.2.0
